requirements.txt
DateTime~=5.1
setuptools~=65.5.1
mongoengine~=0.27.0
oauth2client~=4.1.3
google-api-python-client~=2.82.0
gspread~=5.7.2
requests~=2.28.2
Pillow~=9.4.0
aiohttp~=3.8.4